src/requirements.txt
fastapi
uvicorn

opentelemetry-distro
opentelemetry-exporter-otlp

python-dotenv